+++ frontend/src/pages/templates/Default.template.jsx
@@ -1,0 +1,17 @@
+import { Outlet } from "react-router-dom";
+import Headroom from "react-headroom";
+import Navbar from "../../components/Navbar";
+import { SearchContextProvider } from "../../contexts/SearchContext";
+
+function DefaultTemplate() {
+  return (
+    <SearchContextProvider>
+      <Headroom>
+        <Navbar />
+      </Headroom>
+      <Outlet />
+    </SearchContextProvider>
+  );
+}
+
+export default DefaultTemplate;
